Add request_feedback cli command, some cleanup
