+++ Procfile
@@ -1,1 +1,1 @@
-web: waitress-serve --port=${PORT:-5000} --threads=${WEB_CONCURRENCY:-4} stanley.main:app
+web: waitress-serve --port=${PORT:-5000} --threads=${WEB_CONCURRENCY:-4} stanley.app:app
